refactored user-settings directory and fixed user property
