+++ frontend/src/main/java/com/example/frontend/controllers/SettingsController.java
@@ -11,11 +11,11 @@ import java.io.IOException;
 public class SettingsController {
 
     @FXML
-    private TextField Name_Box; // Change type to TextField
+    private TextField Name_Box;
 
     @FXML
     protected void Return_To_Main_Page() {
-        // Implement your logic here if needed
+        loadScene("/com/example/frontend/setting-page.fxml");
     }
 
     private void loadScene(String fxmlPath) {
@@ -31,22 +31,24 @@ public class SettingsController {
     }
 
     @FXML
-    public void Go_To_Change_First_Name() {
-        loadScene("/com/example/frontend/ChangeFirstName.fxml");
+    public void changeFirstName() {
+        loadScene("/com/example/frontend/userSettings/changeFirstName.fxml");
     }
 
     @FXML
-    public void Go_To_Change_Last_Name() {
-        loadScene("/com/example/frontend/ChangeLastName.fxml");
+    public void changeLastName() {
+        loadScene("/com/example/frontend/userSettings/changeLastName.fxml");
     }
 
     @FXML
-    public void Go_To_Change_Username() {
-        loadScene("/com/example/frontend/ChangeUsername.fxml");
+    public void changeEmail() {
+        loadScene("/com/example/frontend/userSettings/changeEmail.fxml");
     }
 
     @FXML
-    public void Go_To_Change_Email() {
-        loadScene("/com/example/frontend/ChangeEmail.fxml");
+    public void changePassword() {
+        loadScene("/com/example/frontend/userSettings/changePassword.fxml");
     }
+
+
 }
