+++ frontend/src/main/java/com/example/frontend/controllers/user_setting/ChangeFirstNameController.java
@@ -1,4 +1,4 @@
-package com.example.frontend.controllers;
+package com.example.frontend.controllers.user_setting;
 
 import javafx.fxml.FXML;
 import javafx.scene.control.Button;
@@ -8,13 +8,11 @@ import java.io.IOException;
 
 public class ChangeFirstNameController {
 
-    @FXML
-    private Button Back_Button;
+    @FXML private Button backButton;
 
-    @FXML
-    private Button Save_New_First_Name_Button;
+    @FXML private Button saveButton;
 
-    public void Save_New_First_Name() throws IOException {
+    public void saveFirstNameHandler() throws IOException {
     }
 
     public void BackHandler() throws IOException {
