+++ frontend/src/main/resources/com/example/frontend/userSettings/changeEmail.fxml
@@ -6,29 +6,29 @@
 <?import javafx.scene.layout.AnchorPane?>
 <?import javafx.scene.text.Font?>
 
-<AnchorPane prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/19" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.example.frontend.controllers.ChangeUsernameController">
+<AnchorPane prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/19" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.example.frontend.controllers.user_setting.ChangeEmailController">
     <children>
-        <Label layoutX="18.0" layoutY="89.0" text="Current Username:">
+        <Label layoutX="18.0" layoutY="89.0" text="Current Email:">
             <font>
                 <Font name="Times New Roman" size="28.0" />
             </font>
         </Label>
-        <Label layoutX="18.0" layoutY="184.0" text="New Username:">
+        <Label layoutX="18.0" layoutY="184.0" text="New Email:">
             <font>
                 <Font name="Times New Roman" size="28.0" />
             </font>
         </Label>
-        <TextField fx:id="Current_Username_Field" layoutX="260.0" layoutY="80.0">
+        <TextField fx:id="Current_Email_Field" layoutX="260.0" layoutY="80.0">
             <font>
                 <Font name="Times New Roman" size="28.0" />
             </font>
         </TextField>
-        <TextField fx:id="New_Username_Field" layoutX="260.0" layoutY="175.0">
+        <TextField fx:id="New_Email_Field" layoutX="260.0" layoutY="175.0">
             <font>
                 <Font name="Times New Roman" size="28.0" />
             </font>
         </TextField>
-        <Button fx:id="Save_New_Username_Button" layoutX="259.0" layoutY="263.0" mnemonicParsing="false" onAction="#Save_New_Username" text="Save">
+        <Button fx:id="Save_New_Email_Button" layoutX="259.0" layoutY="263.0" mnemonicParsing="false" onAction="#saveEmailHandler" text="Save">
             <font>
                 <Font name="Times New Roman" size="28.0" />
             </font>
